Move query validator to separate file
